rest-gateway/requirements.txt
Django==4.1.6
djangorestframework==3.14.0
requests==2.28.2